+++ system/src/migrations/updates/Update_240930_1.mjs
@@ -1,0 +1,16 @@
+import { FalloutUpdateBase } from "../FalloutUpdateBase.mjs";
+
+export default class Update_240930_1 extends FalloutUpdateBase {
+
+	static version = 240930.1;
+
+	async updateItem(itemData, actorData) {
+		if (itemData.type !== "robot_armor") return;
+
+		const updateData = {
+			"system.-=mods": null,
+		};
+
+		return updateData;
+	}
+}