+++ system/src/sheets/_module.mjs
@@ -4,3 +4,4 @@ export {default as FalloutNpcSheet} from "./FalloutNpcSheet.mjs";
 export {default as FalloutPcSheet} from "./FalloutPcSheet.mjs";
 export {default as FalloutScavengingLocationSheet} from "./FalloutScavengingLocationSheet.mjs";
 export {default as FalloutSettlementSheet} from "./FalloutSettlementSheet.mjs";
+export {default as FalloutVehicleSheet } from "./FalloutVehicleSheet.mjs";